Move wpAttachments components
