+++ frontend/app/components/wp-attachments/wp-attachments.directive.html
